Move ratelimiter to authentication
